chore: refactor for filter-packages to be part of the core again
