+++ packages/core/package.json
@@ -38,7 +38,6 @@
     "@inquirer/expand": "^4.0.0",
     "@inquirer/input": "^4.0.0",
     "@inquirer/select": "^4.0.0",
-    "@lerna-lite/filter-packages": "workspace:*",
     "@lerna-lite/npmlog": "workspace:*",
     "@npmcli/run-script": "^8.1.0",
     "clone-deep": "^4.0.1",
@@ -52,6 +51,7 @@
     "json5": "^2.2.3",
     "load-json-file": "^7.0.1",
     "minimatch": "^9.0.5",
+    "multimatch": "^7.0.0",
     "npm-package-arg": "^11.0.3",
     "p-map": "^7.0.2",
     "p-queue": "^8.0.1",
