+++ packages/core/src/filter-packages/__fixtures__/filtering/lerna.json
