+++ packages/core/src/filter-packages/__fixtures__/filtering/package.json
