+++ packages/core/src/filter-packages/__fixtures__/filtering/packages/package-1/package.json
