+++ packages/core/src/filter-packages/__fixtures__/filtering/packages/package-2/package.json
