+++ packages/core/src/filter-packages/__fixtures__/filtering/packages/package-3/package.json
