+++ packages/core/src/filter-packages/__fixtures__/filtering/packages/package-4/package.json
