+++ packages/core/src/filter-packages/__fixtures__/filtering/packages/package-5/package.json
