+++ packages/core/src/filter-packages/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/filter-packages.js';
+export * from './lib/get-filtered-packages.js';
+export * from './lib/interfaces.js';
