+++ packages/core/src/filter-packages/lib/filter-packages.ts
