+++ packages/core/src/filter-packages/lib/get-filtered-packages.ts
