+++ packages/core/src/index.ts
@@ -1,6 +1,7 @@
 // folders
 export * from './corepack/exec-package-manager.js';
 export * from './corepack/is-corepack-enabled.js';
+export * from './filter-packages/index.js';
 export * from './models/command-options.js';
 export * from './models/interfaces.js';
 export * from './package-graph/package-graph.js';
@@ -38,6 +39,7 @@ export * from './utils/string-utils.js';
 export * from './utils/types.js';
 export * from './utils/warn-if-hanging.js';
 export * from './utils/write-log-file.js';
+
 
 // files
 export * from './child-process.js';
