+++ packages/exec/src/exec-command.ts
@@ -3,6 +3,8 @@ import {
   Command,
   CommandType,
   ExecCommandOption,
+  FilterOptions,
+  getFilteredPackages,
   logOutput,
   Package,
   ProjectConfig,
@@ -11,7 +13,6 @@ import {
   spawnStreaming,
   ValidationError,
 } from '@lerna-lite/core';
-import { FilterOptions, getFilteredPackages } from '@lerna-lite/filter-packages';
 import { Profiler } from '@lerna-lite/profiler';
 import pMap from 'p-map';
 import c from 'tinyrainbow';
