--- packages/filter-packages/.npmignore
@@ -1,2 +1,0 @@
-tsconfig.json
-tsconfig.tsbuildinfo
