--- packages/filter-packages/package.json
@@ -1,46 +1,0 @@
-{
-  "name": "@lerna-lite/filter-packages",
-  "description": "Lerna-Lite common filtering utils used by a few optional commands (Exec/List/Run/Watch)",
-  "version": "3.9.3",
-  "files": [
-    "/dist"
-  ],
-  "type": "module",
-  "publishConfig": {
-    "access": "public"
-  },
-  "types": "./dist/index.d.ts",
-  "exports": {
-    "types": "./dist/index.d.ts",
-    "import": "./dist/index.js"
-  },
-  "scripts": {
-    "prebuild": "rimraf dist",
-    "build": "tsc",
-    "build:incremental": "tsc --incremental --sourceMap false --excludeDirectories dist",
-    "pack-tarball": "npm pack"
-  },
-  "license": "MIT",
-  "author": "Ghislain B.",
-  "homepage": "https://github.com/lerna-lite/lerna-lite",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/lerna-lite/lerna-lite.git",
-    "directory": "packages/filter-packages"
-  },
-  "bugs": {
-    "url": "https://github.com/lerna-lite/lerna-lite/issues"
-  },
-  "engines": {
-    "node": "^18.0.0 || >=20.0.0"
-  },
-  "dependencies": {
-    "@lerna-lite/core": "workspace:*",
-    "@lerna-lite/npmlog": "workspace:*",
-    "multimatch": "^7.0.0"
-  },
-  "devDependencies": {
-    "@types/node": "^22.7.5",
-    "yargs": "^17.7.2"
-  }
-}
