--- packages/filter-packages/src/index.ts
@@ -1,3 +1,0 @@
-export * from './filter-packages.js';
-export * from './get-filtered-packages.js';
-export * from './interfaces.js';
