--- packages/filter-packages/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "extends": "../../tsconfig.base.json",
-  "compileOnSave": false,
-  "compilerOptions": {
-    "rootDir": "src",
-    "outDir": "dist",
-    "types": ["node"]
-  },
-  "filesGlob": ["./src/**/*.ts"],
-  "include": ["src/**/*.ts"],
-  "exclude": ["dist", "node_modules", "**/__mocks__/**", "**/__tests__/**"],
-  "references": [{ "path": "../core" }, { "path": "../npmlog" }]
-}
