+++ packages/list/package.json
@@ -37,7 +37,6 @@
   "dependencies": {
     "@lerna-lite/cli": "workspace:*",
     "@lerna-lite/core": "workspace:*",
-    "@lerna-lite/filter-packages": "workspace:*",
     "@lerna-lite/listable": "workspace:*"
   },
   "devDependencies": {
