+++ packages/list/src/list-command.ts
@@ -1,5 +1,4 @@
-import { Command, CommandType, ListCommandOption, logOutput, ProjectConfig } from '@lerna-lite/core';
-import { FilterOptions, getFilteredPackages } from '@lerna-lite/filter-packages';
+import { Command, CommandType, FilterOptions, ListCommandOption, getFilteredPackages, logOutput, ProjectConfig } from '@lerna-lite/core';
 import { listable } from '@lerna-lite/listable';
 
 export function factory(argv: ListCommandOption) {
