+++ packages/run/README.md
@@ -50,7 +50,7 @@ $ lerna run build \$LERNA_PACKAGE_NAME
 
 ## Options
 
-`lerna run` accepts all [filter flags](https://github.com/lerna-lite/lerna-lite/blob/main/packages/filter-packages/README.md#options).
+`lerna run` accepts all [filter flags](https://github.com/lerna-lite/lerna-lite/blob/main/packages/core/README.md#options).
 
 ```sh
 $ lerna run --scope my-component test
